+++ src/importlinter/domain/fields.py
@@ -18,6 +18,10 @@ class ValidationError(Exception):
         self.message = message
 
 
+class NotParsedError(Exception):
+    pass
+
+
 class Field(Generic[FieldValue], abc.ABC):
     """
     Base class for containers for some data on a Contract.
@@ -54,30 +58,34 @@ class Field(Generic[FieldValue], abc.ABC):
     def parse(self, raw_data: Union[str, List[str]]) -> FieldValue:
         """
         Given some raw data supplied by a user, return some clean data.
-
-        Raises:
-            ValidationError if the data is invalid.
         """
         raise NotImplementedError
 
+    @property
+    def value(self) -> FieldValue:
+        raise NotParsedError(
+            "Field is still a class-level descriptor. It has not been parsed yet or "
+            "it does not support value property access."
+        )
 
-class StringField(Field):
+
+class StringField(Field[str]):
     """
     A field for single values of strings.
     """
 
-    def parse(self, raw_data: Union[str, List]) -> str:
+    def parse(self, raw_data: Union[str, List[str]]) -> str:
         if isinstance(raw_data, list):
             raise ValidationError("Expected a single value, got multiple values.")
         return str(raw_data)
 
 
-class BooleanField(Field):
+class BooleanField(Field[bool]):
     """
     A field for single values of booleans.
     """
 
-    def parse(self, raw_data: Union[str, List]) -> bool:
+    def parse(self, raw_data: Union[str, List[str]]) -> bool:
         if isinstance(raw_data, list):
             raise ValidationError("Expected a single value, got multiple values.")
 
@@ -87,6 +95,17 @@ class BooleanField(Field):
             return False
         else:
             raise ValidationError(f"Could not parse a boolean from '{raw_data}'.")
+
+
+class IntegerField(Field[int]):
+    """
+    A field for single values of integers.
+    """
+
+    def parse(self, raw_data: Union[str, List[str]]) -> int:
+        if isinstance(raw_data, list):
+            raise ValidationError("Expected a single value, got multiple values.")
+        return int(raw_data)
 
 
 class BaseMultipleValueField(Field):