+++ tests/helpers/contracts.py
@@ -88,12 +88,12 @@ class MutationCheckContract(Contract):
     will fail, if the graph gets mutated by other contracts.
     """
 
-    number_of_modules = fields.StringField()
-    number_of_imports = fields.StringField()
+    number_of_modules = fields.IntegerField()
+    number_of_imports = fields.IntegerField()
 
     def check(self, graph: ImportGraph, verbose: bool) -> ContractCheck:
-        number_of_modules: int = int(self.number_of_modules)  # type: ignore
-        number_of_imports: int = int(self.number_of_imports)  # type: ignore
+        number_of_modules = self.number_of_modules.value
+        number_of_imports = self.number_of_imports.value
         if not all(
             [
                 number_of_modules == len(graph.modules),